Split 'welcome' feature actions and reducers into 'connection' and 'conference' for better separation of concerns
