+++ features/welcome/index.js
@@ -1,3 +1,1 @@
-export * from './actions';
-export * from './actionTypes';
 export * from './components';
